pyproject.toml
[project]
name = "doclify"
version = "0.0.0.1"
dependencies = [
    "click",
    "google-genai",
    "pathspec>=1.0.3",
    "pyyaml",
    "rich",
    "twine>=6.2.0",
]

[project.scripts]
doclify = "doclify.pipelines.supervisor:cli"

[tool.setuptools]
packages = ["doclify"]

[tool.setuptools.package-data]
"doclify" = ["resources/*.gitignore"]
